remove global_defs
