+++ pkg/store/context.go
@@ -27,7 +27,6 @@ type Stores struct {
 	Defs               Defs
 	Directory          Directory
 	ExternalAuthTokens ExternalAuthTokens
-	GlobalDefs         GlobalDefs
 	GlobalDeps         GlobalDeps
 	GlobalRefs         GlobalRefs
 	Graph              srcstore.MultiRepoStoreImporterIndexer
@@ -52,7 +51,6 @@ const (
 	_DefsKey
 	_DirectoryKey
 	_ExternalAuthTokensKey
-	_GlobalDefsKey
 	_GlobalDepsKey
 	_GlobalRefsKey
 	_GraphKey
@@ -91,9 +89,6 @@ func WithStores(ctx context.Context, s Stores) context.Context {
 	}
 	if s.ExternalAuthTokens != nil {
 		ctx = WithExternalAuthTokens(ctx, s.ExternalAuthTokens)
-	}
-	if s.GlobalDefs != nil {
-		ctx = WithGlobalDefs(ctx, s.GlobalDefs)
 	}
 	if s.GlobalDeps != nil {
 		ctx = WithGlobalDeps(ctx, s.GlobalDeps)
@@ -239,20 +234,6 @@ func ExternalAuthTokensFromContext(ctx context.Context) ExternalAuthTokens {
 	s, ok := ctx.Value(_ExternalAuthTokensKey).(ExternalAuthTokens)
 	if !ok || s == nil {
 		panic("no ExternalAuthTokens set in context")
-	}
-	return s
-}
-
-// WithGlobalDefs returns a copy of parent with the given GlobalDefs store.
-func WithGlobalDefs(parent context.Context, s GlobalDefs) context.Context {
-	return context.WithValue(parent, _GlobalDefsKey, s)
-}
-
-// GlobalDefsFromContext gets the context's GlobalDefs store. If the store is not present, it panics.
-func GlobalDefsFromContext(ctx context.Context) GlobalDefs {
-	s, ok := ctx.Value(_GlobalDefsKey).(GlobalDefs)
-	if !ok || s == nil {
-		panic("no GlobalDefs set in context")
 	}
 	return s
 }
