+++ pkg/store/global_graph.go
@@ -7,26 +7,6 @@ import (
 	"sourcegraph.com/sourcegraph/srclib/unit"
 )
 
-// GlobalDefs defines the interface for searching global defs.
-type GlobalDefs interface {
-	// Search performs a global search for defs that match the given repo, unit and def
-	// query, and ranks the defs by a combination of bag of words similarity and global ref count.
-	Search(ctx context.Context, op *GlobalDefSearchOp) (*sourcegraph.SearchResultsList, error)
-
-	// Update takes the graph output of a list of source units and
-	// updates the set of defs in the global def store that originate
-	// from these source units. If a repository is specified in op
-	// with an empty source unit, then Update updates the data for
-	// the entire repository.
-	Update(ctx context.Context, op GlobalDefUpdateOp) error
-
-	// RefreshRefCounts computes and sets the global ref counts of all defs in the
-	// specified source units. If a repository is specified in op with
-	// an empty source unit, then RefreshRefCounts updates ref counts
-	// for the entire repository.
-	RefreshRefCounts(ctx context.Context, op GlobalDefUpdateOp) error
-}
-
 // GlobalRefs defines the interface for getting and listing global ref locations.
 type GlobalRefs interface {
 	// Get returns the names and ref counts of all repos and files within those repos
@@ -43,21 +23,6 @@ type GlobalRefs interface {
 type DefExamples interface {
 	// Get returns the list of example locations for a given def.
 	Get(ctx context.Context, op *sourcegraph.DefsListExamplesOp) (*sourcegraph.RefLocationsList, error)
-}
-
-type GlobalDefSearchOp struct {
-	UnitQuery     string
-	UnitTypeQuery string
-
-	// TokQuery is a list of tokens that describe the user's text
-	// query. Order matter, as the last token is given especial weight.
-	TokQuery []string
-
-	Opt *sourcegraph.SearchOptions
-}
-
-type GlobalDefUpdateOp struct {
-	RepoUnits []RepoUnit
 }
 
 type RepoUnit struct {
