+++ pkg/store/mockstore/global_graph_mock.go
@@ -9,26 +9,6 @@ import (
 	"sourcegraph.com/sourcegraph/srclib/unit"
 )
 
-type GlobalDefs struct {
-	Search_           func(ctx context.Context, op *store.GlobalDefSearchOp) (*sourcegraph.SearchResultsList, error)
-	Update_           func(ctx context.Context, op store.GlobalDefUpdateOp) error
-	RefreshRefCounts_ func(ctx context.Context, op store.GlobalDefUpdateOp) error
-}
-
-func (s *GlobalDefs) Search(ctx context.Context, op *store.GlobalDefSearchOp) (*sourcegraph.SearchResultsList, error) {
-	return s.Search_(ctx, op)
-}
-
-func (s *GlobalDefs) Update(ctx context.Context, op store.GlobalDefUpdateOp) error {
-	return s.Update_(ctx, op)
-}
-
-func (s *GlobalDefs) RefreshRefCounts(ctx context.Context, op store.GlobalDefUpdateOp) error {
-	return s.RefreshRefCounts_(ctx, op)
-}
-
-var _ store.GlobalDefs = (*GlobalDefs)(nil)
-
 type GlobalRefs struct {
 	Get_    func(ctx context.Context, op *sourcegraph.DefsListRefLocationsOp) (*sourcegraph.RefLocationsList, error)
 	Update_ func(ctx context.Context, op *sourcegraph.DefsRefreshIndexOp) error
