+++ pkg/store/mockstore/mockstores.go
@@ -26,7 +26,6 @@ type Stores struct {
 	Defs               Defs
 	Directory          Directory
 	ExternalAuthTokens ExternalAuthTokens
-	GlobalDefs         GlobalDefs
 	GlobalDeps         GlobalDeps
 	GlobalRefs         GlobalRefs
 	Graph              srcstore.MockMultiRepoStore
@@ -50,7 +49,6 @@ func (s *Stores) Stores() store.Stores {
 		Defs:               &s.Defs,
 		Directory:          &s.Directory,
 		ExternalAuthTokens: &s.ExternalAuthTokens,
-		GlobalDefs:         &s.GlobalDefs,
 		GlobalDeps:         &s.GlobalDeps,
 		GlobalRefs:         &s.GlobalRefs,
 		Graph:              &s.Graph,
