+++ services/backend/internal/middleware/outer/outer_middleware.go
@@ -2182,36 +2182,6 @@ func (s wrappedSearch) Search(ctx context.Context, v1 *sourcegraph.SearchOp) (re
 	return rv, nil
 }
 
-func (s wrappedSearch) RefreshIndex(ctx context.Context, v1 *sourcegraph.SearchRefreshIndexOp) (returnedResult *pbtypes.Void, returnedError error) {
-	defer func() {
-		if err := recover(); err != nil {
-			const size = 64 << 10
-			buf := make([]byte, size)
-			buf = buf[:runtime.Stack(buf, false)]
-			returnedError = grpc.Errorf(codes.Internal, "panic in Search.RefreshIndex: %v\n\n%s", err, buf)
-			returnedResult = nil
-		}
-	}()
-
-	var err error
-	ctx, err = initContext(ctx, s.ctxFunc, s.services)
-	if err != nil {
-		return nil, wrapErr(err)
-	}
-
-	innerSvc := svc.SearchOrNil(ctx)
-	if innerSvc == nil {
-		return nil, grpc.Errorf(codes.Unimplemented, "Search")
-	}
-
-	rv, err := innerSvc.RefreshIndex(ctx, v1)
-	if err != nil {
-		return nil, wrapErr(err)
-	}
-
-	return rv, nil
-}
-
 type wrappedUsers struct {
 	ctxFunc  ContextFunc
 	services svc.Services
